+++ src/Client/Window.cpp
@@ -3,16 +3,6 @@
 
 void Window::drawImGui()
 {
-    #ifndef GEODE_IS_MACOS
-    if (auto action = typeinfo_cast<CCActionInterval*>(getActionByTag(69)))
-    {
-        action->step(CCDirector::get()->getDeltaTime());
-
-        if (action->isDone())
-            this->stopAction(action);
-    }
-    #endif
-
     ImGui::SetNextWindowPos(ImVec2(getPosition().x, getPosition().y));
     ImGui::SetNextWindowSize(getDesiredWindowSize());